volleyball_tracking/train/requirements.txt
# 训练环境依赖
# Python 3.8+

# 深度学习框架
torch==2.1.0
torchvision==0.16.0

# YOLO
ultralytics==8.1.0

# 数据处理
numpy==1.24.3
opencv-python==4.8.1.78
scipy==1.11.3
Pillow==10.1.0

# 数据增强
albumentations==1.3.1

# 可视化
matplotlib==3.8.0
seaborn==0.13.0

# 训练监控
tensorboard==2.15.1
tqdm==4.66.1

# 工具
pyyaml==6.0.1
pandas==2.1.3
